refactor: requestContext used in middlewares
Also refactored handlers to separate into 2 files main cert-related routes and helper routes. In this process decided to completely remove use of middlewares for helper routes. They are supposed to be simple anyways, so can handle their own logging/errors without too much clutter.

The middlewares are currently hardcoded to only work for cert routes. We could try to expand them and make them work for all routes in the future, but its not even clear how much benefit we get out of that.
