+++ server/middleware/logging.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/Layr-Labs/eigenda-proxy/common/types/commitments"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 )
 
@@ -11,9 +12,10 @@ import (
 // It does not write anything to the response, that is the job of the handlers.
 // Currently we cannot log the status code because go's default ResponseWriter interface does not expose it.
 // TODO: implement a ResponseWriter wrapper that saves the status code: see https://github.com/golang/go/issues/18997
-func WithLogging(
+func withLogging(
 	handleFn func(http.ResponseWriter, *http.Request) error,
 	log logging.Logger,
+	mode commitments.CommitmentMode,
 ) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		start := time.Now()
@@ -24,13 +26,7 @@ func WithLogging(
 		args := []any{
 			"method", r.Method, "url", r.URL,
 			"status", scw.status, "duration", time.Since(start),
-		}
-
-		ctx := GetRequestContext(r)
-		if ctx == nil {
-			log.Error("logging middleware: request context not found")
-		} else {
-			args = append(args, []any{"commitment_mode", ctx.CommitmentMode, "cert_version", ctx.CertVersion})
+			"commitment_mode", mode, "cert_version", getCertVersion(r),
 		}
 
 		if err != nil {
